Move collector functions out of headers
