+++ projects/rocprofiler-compute/src/lib/torch_trace_collector/leaf_context.h
@@ -15,7 +15,7 @@ inline constexpr const char* kAutogradBackwardLeaf = "#1@autograd.bwd:0";
 
 // Returns the default leaf-context label emitted by the producer for a
 // given scope. The Python coverage parser matches these tokens exactly.
-inline const char* default_leaf_context(bool is_backward_scope, std::int64_t seq_nr, bool stack_was_empty)
+constexpr const char* default_leaf_context(bool is_backward_scope, std::int64_t seq_nr, bool stack_was_empty)
 {
     if (is_backward_scope)
     {
